LAB06 Tests to do;

+++ src/main/java/pl/edu/pjatk/MPR_Project/service/MyRestService.java
@@ -2,17 +2,20 @@ package pl.edu.pjatk.MPR_Project.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import pl.edu.pjatk.MPR_Project.exception.CapybaraAlreadyExists;
+import pl.edu.pjatk.MPR_Project.exception.CapybaraNotFoundException;
+import pl.edu.pjatk.MPR_Project.exception.InvalidInputCapybaraException;
 import pl.edu.pjatk.MPR_Project.model.Capybara;
 import pl.edu.pjatk.MPR_Project.repository.CapybaraRepository;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class MyRestService {
-    private CapybaraRepository repository;
-    private StringService stringService;
+    private final CapybaraRepository capybaraRepository;
+    private final CapybaraRepository repository;
+    private final StringService stringService;
 
     @Autowired
     public MyRestService(CapybaraRepository repository, StringService stringService) {
@@ -23,39 +26,62 @@ public class MyRestService {
         Capybara capy2 = new Capybara("Capy2", 3);
         Capybara capy3 = new Capybara("Capy3", 4);
 
-        capy1.setIdentification(capy1);
-        capy2.setIdentification(capy2);
-        capy3.setIdentification(capy3);
+        capy1.setIdentification();
+        capy2.setIdentification();
+        capy3.setIdentification();
 
         repository.save(capy1);
         repository.save(capy2);
         repository.save(capy3);
+        this.capybaraRepository = repository;
     }
 
     public void addCapybara(Capybara capybara) {
         capybara.setName(stringService.uppercase(capybara.getName()));
 
-        capybara.setIdentification(capybara);
+        capybara.setIdentification();
+
+        long identification = capybara.getIdentification();
+
+        if(capybara.getName().isBlank() || capybara.getAge() == 0) {
+            throw new InvalidInputCapybaraException();
+        }
+
+        if(capybaraRepository.findByIdentification(identification).isPresent()) {
+            throw new CapybaraAlreadyExists();
+        }
+
         this.repository.save(capybara);
     }
 
     public void patchCapybaraById(Capybara capybara, Long id) {
-       Optional<Capybara> capybaraToReplaceOptional = repository.findById(id);
+        Optional<Capybara> capybaraToReplaceOptional = repository.findById(id);
 
-        if(capybaraToReplaceOptional.isPresent()) {
-            Capybara capybaraToReplace = capybaraToReplaceOptional.get();
+        if (capybaraToReplaceOptional.isEmpty()) {
+            throw new CapybaraNotFoundException();
+        }
+        if(capybaraToReplaceOptional.get().getName().isEmpty() || capybaraToReplaceOptional.get().getName() == null || capybaraToReplaceOptional.get().getAge() == 0) {
+            throw new InvalidInputCapybaraException();
+        }
+        Capybara capybaraToReplace = capybaraToReplaceOptional.get();
 
-            capybaraToReplace.setName(stringService.uppercase(capybara.getName()));
-            capybaraToReplace.setAge(capybara.getAge());
+        capybaraToReplace.setName(stringService.uppercase(capybara.getName()));
+        capybaraToReplace.setAge(capybara.getAge());
 
-            capybaraToReplace.setIdentification(capybaraToReplace);
+        capybaraToReplace.setIdentification();
 
-            repository.save(capybaraToReplace);
+        long identification = capybaraToReplace.getIdentification();
+        if(capybaraRepository.findByIdentification(identification).isPresent()) {
+            throw new CapybaraAlreadyExists();
         }
 
+        repository.save(capybaraToReplace);
     }
 
     public void deleteCapybaraById(Long id) {
+        if (repository.findById(id).isEmpty()) {
+            throw new CapybaraNotFoundException();
+        }
         repository.deleteById(id);
     }
 
@@ -63,6 +89,9 @@ public class MyRestService {
         List<Capybara> capybaraListToLower = repository.findByName(name);
         capybaraListToLower.forEach(capybara -> capybara.setName(stringService.lowercase(capybara.getName())));
 
+        if (capybaraListToLower.isEmpty()) {
+            throw new CapybaraNotFoundException();
+        }
         return capybaraListToLower;
     }
 
@@ -70,21 +99,31 @@ public class MyRestService {
         List<Capybara> capybaraListToLower = repository.findByAge(age);
         capybaraListToLower.forEach(capybara -> capybara.setName(stringService.lowercase(capybara.getName())));
 
+        if (capybaraListToLower.isEmpty()) {
+            throw new CapybaraNotFoundException();
+        }
         return capybaraListToLower;
     }
 
-    public Optional<Capybara> getById(Long id) {
+    public Capybara getById(Long id) {
         Optional<Capybara> capybaraToLower = repository.findById(id);
-        capybaraToLower.ifPresent(capybara -> capybara.setName(stringService.lowercase(capybara.getName())));
 
-        return capybaraToLower;
+        if (capybaraToLower.isEmpty()) {
+            throw new CapybaraNotFoundException();
+        }
+        capybaraToLower.get().setName(stringService.lowercase(capybaraToLower.get().getName()));
+
+        return capybaraToLower.get();
     }
 
     public List<Capybara> getAllCapybaraObjects() {
-        List<Capybara> capybaraList = (List<Capybara>) repository.findAll();
-        capybaraList.forEach(capybara -> capybara.setName(stringService.lowercase(capybara.getName())));
+        List<Capybara> capybaraListToLower = (List<Capybara>) repository.findAll();
+        capybaraListToLower.forEach(capybara -> capybara.setName(stringService.lowercase(capybara.getName())));
 
-        return capybaraList;
+        if (capybaraListToLower.isEmpty()) {
+            throw new CapybaraNotFoundException();
+        }
+        return capybaraListToLower;
     }
 }
